Separando os arquios array.h, str.h, stg.h, model.h, pad.h.
Falta corrigir possíveis bugs
