+++ include/array/len.h
@@ -1,0 +1,6 @@
+#ifndef S3D_ARRAY_LEN_H
+#define S3D_ARRAY_LEN_H
+
+#define s3d_array_len(arr) (sizeof(arr) / sizeof((arr)[0]))
+
+#endif
