+++ include/model/idxtab.h
@@ -1,0 +1,10 @@
+#ifndef S3D_MODEL_IDXTAB_H
+#define S3D_MODEL_IDXTAB_H
+
+typedef struct {
+    int num_rows;
+
+    s3d_model_idxtab_row *rows;
+} s3d_model_idxtab;
+
+#endif
