+++ include/model/idxtab_row.h
@@ -1,0 +1,9 @@
+#ifndef S3D_MODEL_IDXTAB_ROW_H
+#define S3D_MODEL_IDXTAB_ROW_H
+
+typedef struct {
+    void *idx_ptr;
+    int num_idx;
+} s3d_model_idxtab_row;
+
+#endif
