+++ include/pad/type.h
@@ -1,0 +1,9 @@
+#ifndef S3D_PAD_TYPE_H
+#define S3D_PAD_TYPE_H
+
+
+struct {
+    int btns[s3d_pad_num_btns];
+} s3d_pad;
+
+#endif
