+++ include/stage/obj_type.h
@@ -1,0 +1,13 @@
+#ifndef S3D_STAGE_OBJ_TYPE_H
+#define S3D_STAGE_OBJ_TYPE_H
+ 
+enum {
+    s3d_stg_obj_type_none,
+    s3d_stg_obj_type_player,
+    s3d_stg_obj_type_outer_wall,
+    s3d_stg_obj_type_wall,
+    s3d_stg_obj_type_block,
+    s3d_stg_obj_type_slot
+};
+
+#endif
