+++ include/string/arrcat.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_ARRCAT_H
+#define S3D_STRING_ARRCAT_H
+
+#define s3d_strarrcat(dst, src) strlcat(dst, src, sizeof(dst))
+
+#endif
