+++ include/string/arrcpy.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_ARRCPY_H
+#define S3D_STRING_ARRCPY_H
+
+#define s3d_strarrcpy(dst, src) strlcpy(dst, src, sizeof(dst))
+
+#endif
