+++ include/string/charin.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_CHARIN_H
+#define S3D_STRING_CHARIN_H
+
+int s3d_strcharin(char c, const char *chars);
+
+#endif
