+++ include/string/chrcount.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_CHRCOUNT_H
+#define S3D_STRING_CHRCOUNT_H
+
+int s3d_strchrcount(const char *s, int c);
+
+#endif
