+++ include/string/consume.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_CONSUME_H
+#define S3D_STRING_CONSUME_H
+
+int s3d_strconsume(char ** const s, const char *token);
+
+#endif
