+++ include/string/empty.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_EMPTY_H
+#define S3D_STRING_EMPTY_H
+
+#define s3d_strempty(s) ((s)[0] == 0)
+
+#endif
