+++ include/string/endswith.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_ENDSWITH_H
+#define S3D_STRING_ENDSWITH_H
+
+int s3d_strendswith(const char *s, const char *suffix);
+
+#endif
