+++ include/string/rtrim.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_RTRIM_H
+#define S3D_STRING_RTRIM_H
+
+char *s3d_strrtrim(char *s, const char *suffix);
+
+#endif
