+++ include/string/skip.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_SKIP_H
+#define S3D_STRING_SKIP_H
+
+char *s3d_strskip(char ** const s, const char *chars);
+
+#endif
