+++ include/string/startswith.h
@@ -1,0 +1,6 @@
+#ifndef S3D_STRING_STARTSWITH_H
+#define S3D_STRING_STARTSWITH_H
+
+int s3d_strstartswith(const char *s, const char *prefix);
+
+#endif
