--- include/sys.h
@@ -1,7 +1,0 @@
-#ifndef S3D_FILE_H
-#define S3D_FILE_H
-
-char *s3d_read_file(const char *path, int text, int *size);
-void s3d_fsleep(float seconds);
-
-#endif
