+++ src/model/bind_tex.c
@@ -1,0 +1,5 @@
+#include "model/bind_tex.h"
+ 
+void s3d_model_bind_tex(s3d_model *m) {
+    glBindTexture(GL_TEXTURE_2D, m->tex_id);
+}
