+++ src/model/bind_va.c
@@ -1,0 +1,5 @@
+#include "model/bind_va.h"
+
+void s3d_model_bind_va(s3d_model *m) {
+    glBindVertexArray(m->va_id);
+}
