+++ src/pad/callback.c
@@ -1,6 +1,4 @@
-#include "pad.h"
-
-#include<GLFW/glfw3.h>
+#include "pad/callback.h"
 
 void s3d_pad_callback(GLFWwindow *wnd, int key, int scan, int action, int mods) {
     if(action == GLFW_REPEAT) {
@@ -39,28 +37,3 @@ void s3d_pad_callback(GLFWwindow *wnd, int key, int scan, int action, int mods) 
         #undef key_case
     }
 }
-
-void s3d_pad_update() {
-    #define update_btn(btn) \
-        if(s3d_pad_btn_st(btn, hit)) { \
-            s3d_pad_btn(btn) ^= s3d_pad_btn_hit; \
-        } \
-        \
-        if(s3d_pad_btn_st(btn, release)) { \
-            s3d_pad_btn(btn) ^= s3d_pad_btn_release; \
-        }
-
-    update_btn(A);
-    update_btn(X);
-
-    update_btn(L);
-    update_btn(R);
-
-    update_btn(DL);
-    update_btn(DU);
-    update_btn(DR);
-    update_btn(DD);
-
-    #undef update_btn
-}
-
