--- src/sys.c
@@ -1,6 +1,0 @@
-#include "sys.h"
-
-#include<stdio.h>
-#include<unistd.h>
-
-#include "err.h"
